Fix name normalization bug
Group names were not being normalized when loaded from includes. As a
result, they didn't match the normalized names used for the loaded
groups.

This updates the internal name usage to be normalized during
resolution. Errors will still contain the non-normalized name which
the user used if possible (see: the `CyclicDependencyError`
parameters).

A new test confirms the fix, and fails without it.